separate admin crud panel to allow dropdown
