+++ app/Admin/page.tsx
@@ -4,12 +4,11 @@ import { useState } from 'react';
 import PageTitle from '@/components/layout/pageTitle';
 import AdminTabs from '@/components/admin/SiteAdminTabs';
 import AdminCrudPanel from '@/components/admin/AdminCrudPanel';
-import { crudConfigs } from '@/lib/crudConfig';
-import Toast from '@/components/shared/Toast';
-import type { ToastProps } from '@/types/toast';
+import DropdownEditor from '@/components/admin/DropdownEditor';
+import { adminConfigs } from '@/lib/adminConfigs';
 
-// Get all the keys (tab names) from the crudConfigs object
-const tabNames = Object.keys(crudConfigs) as (keyof typeof crudConfigs)[];
+// Get all the keys (tab names) from the adminConfigs object
+const tabNames = Object.keys(adminConfigs) as (keyof typeof adminConfigs)[];
 
 export default function AdminPage() {
   // Set up state to track which tab is currently selected
@@ -17,23 +16,9 @@ export default function AdminPage() {
     tabNames[0]
   );
 
-  // Set up state to show a toast message (or nothing if null)
-  const [toast, setToast] = useState<Omit<ToastProps, 'onClose' | 'duration'> | null>(null);
-
   return (
     <>
       <PageTitle title="SITE ADMIN" />
-
-      {toast && (
-        <div className="fixed top-6 right-6 z-50 animate-fade-in-out">
-          <Toast 
-            type={toast.type}
-            title={toast.title}
-            message={toast.message}
-            onClose={() => setToast(null)} 
-          />
-        </div>
-      )}
       <div className="flex flex-col lg:flex-row px-4 sm:px-6 py-6 sm:py-10 gap-4 sm:gap-6">
         <AdminTabs
           tabs={tabNames}
@@ -41,11 +26,10 @@ export default function AdminPage() {
           setActiveTab={setActiveTab}
         />
         <div className="flex-1">
-          <AdminCrudPanel
-            title={activeTab}
-            config={crudConfigs[activeTab] as any}
-            setToast={setToast}
-          />
+          {adminConfigs[activeTab].component === 'dropdown'
+            ? <DropdownEditor config={adminConfigs[activeTab] as any} />
+            : <AdminCrudPanel title={activeTab} config={adminConfigs[activeTab] as any} />
+          }
         </div>
       </div>
     </>
