+++ components/admin/DropdownEditor.tsx
@@ -1,0 +1,3 @@
+export default function DropdownEditor({ config }: { config: any }) {
+  return <div>Dropdown Editor Coming Soon</div>;
+}
